+++ app/src/main/cpp/widget/LoadingView.cpp
@@ -24,13 +24,15 @@ void LoadingView::draw(SkCanvas *canvas) {
         if (time > endTime || time < startTime) {
             rectHeight = loadingHeight;
         } else {
-            rectHeight = std::min(abs(time - startTime), abs(time - endTime)) * loadingHeight * 2 /
-                         (middleTime - startTime) + loadingHeight;
+            rectHeight = static_cast<int>(
+                    std::min(abs(time - startTime), abs(time - endTime)) * loadingHeight * 2 /
+                    (middleTime - startTime) + loadingHeight);
         }
-        rect.setXYWH(left + width / 2 + loadingWidth * distance + loadingMargin * distance,
-                     top + height / 2 - rectHeight / 2,
-                     loadingWidth,
-                     rectHeight);
+        rect.setXYWH(static_cast<float >(left + width / 2 + loadingWidth * distance +
+                                         loadingMargin * distance),
+                     static_cast<float >(top + height / 2 - rectHeight / 2),
+                     static_cast<float >(loadingWidth),
+                     static_cast<float >(rectHeight));
         canvas->save();
         canvas->setMatrix(viewMatrix);
         canvas->drawRect(rect, *paint);
